fix: TEAM_468 implement O_CLOEXEC, signal reset on exec, and u64 blocked_signals
Fixes three medium-priority issues found in kernel audit:

1. O_CLOEXEC Support:
   - Add cloexec field to FdEntry struct
   - Add alloc_cloexec(), close_cloexec(), get_cloexec(), set_cloexec() methods
   - Update sys_openat to pass O_CLOEXEC flag
   - Update sys_fcntl F_GETFD/F_SETFD/F_DUPFD_CLOEXEC to use new methods
   - Call close_cloexec() in execve before execution

2. Signal Handler Reset on exec:
   - Reset non-SIG_IGN handlers to SIG_DFL on exec (POSIX requirement)

3. blocked_signals u32 → u64:
   - Change blocked_signals from AtomicU32 to AtomicU64
   - Update all initialization sites in lib.rs, fork.rs, thread.rs
   - Update sys_sigprocmask to read/write 8 bytes
   - Update signal delivery in main.rs for u64 comparison

+++ levitate/src/main.rs
@@ -87,7 +87,8 @@ mod aarch64_handlers {
         let task = current_task();
 
         // Get pending signals that are not blocked
-        let pending = task.pending_signals.load(Ordering::Acquire);
+        // TEAM_468: Cast pending to u64 since blocked_signals is now 64-bit
+        let pending = task.pending_signals.load(Ordering::Acquire) as u64;
         let blocked = task.blocked_signals.load(Ordering::Acquire);
         let deliverable = pending & !blocked;
